requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
pylibmc==1.2.3
django-heroku-memcacheify==0.5
-r requirements/production.txt